8263564: Consolidate POSIX code for runtime exit support: os::shutdown, os::abort and os::die
